[Dtos] Changing Dto's ids to long

+++ Service.API/Dtos/Fantasy.API.Dtos/Fantasy.API.Dtos/Response/FantasyData/VenueDto.cs
@@ -8,7 +8,7 @@ namespace Fantasy.API.Dtos.Response.FantasyData
 {
     public class VenueDto
     {
-        public int VenueId { get; set; }
+        public long VenueId { get; set; }
 
         public string Name { get; set; }
